Improve doc and move some code-related terms to `.code` package Also, remove `.field` package
