+++ tools/validation-generator/src/main/java/io/spine/tools/validate/ValidationCodeGenerator.java
@@ -32,12 +32,12 @@ import io.spine.option.GoesOption;
 import io.spine.option.IfInvalidOption;
 import io.spine.protobuf.AnyPacker;
 import io.spine.tools.validate.code.BooleanExpression;
+import io.spine.tools.validate.code.Check;
+import io.spine.tools.validate.code.ConstraintCode;
+import io.spine.tools.validate.code.CreateViolation;
 import io.spine.tools.validate.code.Expression;
 import io.spine.tools.validate.code.IsSet;
 import io.spine.tools.validate.code.NewViolation;
-import io.spine.tools.validate.field.Check;
-import io.spine.tools.validate.field.ConstraintCode;
-import io.spine.tools.validate.field.CreateViolation;
 import io.spine.type.MessageType;
 import io.spine.validate.Alternative;
 import io.spine.validate.ComparableNumber;
@@ -65,9 +65,9 @@ import java.util.function.Function;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static io.spine.tools.validate.ValidateMethod.VIOLATIONS;
 import static io.spine.tools.validate.code.BooleanExpression.fromCode;
+import static io.spine.tools.validate.code.Containers.isEmpty;
 import static io.spine.tools.validate.code.IsSet.alternativeIsSet;
 import static io.spine.tools.validate.code.VoidExpression.formatted;
-import static io.spine.tools.validate.field.Containers.isEmpty;
 import static io.spine.util.Exceptions.unsupported;
 import static io.spine.util.Preconditions2.checkNotEmptyOrBlank;
 import static io.spine.validate.diags.ViolationText.errorMessage;
